Completed probability stuff for the Markov class. Should now select a note to play based on the probabilities (not yet decided) and the current note heard. Chord currently doesn't affect outcome.

+++ Project/Source/Analysis/Markov.h
@@ -23,7 +23,7 @@ public:
 
 private:
 
-	std::vector<std::vector<float>> matrix;
+	std::vector<std::vector<double>> matrix;
 };
